pyproject.toml
[project]
name = "imgeda"
dynamic = ["version"]
description = "High-performance image dataset exploratory data analysis CLI tool"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Randall Hunt (Caylent)", email = "randall.hunt@caylent.com" }
]
requires-python = ">=3.10"
keywords = ["image", "eda", "dataset", "analysis", "quality", "duplicates", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "questionary>=2.1",
    "Pillow>=11.0",
    "numpy>=2.0",
    "imagehash>=4.3",
    "orjson>=3.10",
    "matplotlib>=3.10",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/caylent/imgeda"
Repository = "https://github.com/caylent/imgeda"
"Bug Tracker" = "https://github.com/caylent/imgeda/issues"

[project.optional-dependencies]
opencv = ["opencv-python-headless>=4.10"]
parquet = ["pyarrow>=15.0"]
embeddings = ["open-clip-torch>=2.24", "torch>=2.0", "umap-learn>=0.5"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "ruff>=0.9",
    "mypy>=1.14",
    "moto[s3]>=5.0",
]

[project.scripts]
imgeda = "imgeda.cli.app:app"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.build.targets.wheel]
packages = ["src/imgeda"]

[tool.hatch.build.targets.sdist]
exclude = [
    "cdk/",
    "docs/",
    "examples/",
    "data/",
    "plots/",
    "tests/",
    ".claude/",
    ".github/",
    "CLAUDE.md",
    ".python-version",
    "uv.lock",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true

[[tool.mypy.overrides]]
module = ["open_clip.*", "torch.*", "umap.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
markers = [
    "integration: Local integration tests (full CLI pipeline, no AWS)",
    "aws: Tests requiring real AWS credentials",
    "slow: Long-running tests",
]
